# api/openapi.yaml
openapi: 3.0.3
info:
  title: openbot API
  version: 0.1.0
  description: |
    HTTP API for openbotd. Authenticated with a session cookie (`openbot_session`).
    Mutating requests must send `X-Openbot-CSRF: 1` (checked together with the Origin header).

    Realtime updates are delivered over a WebSocket at `GET /api/ws`; each frame is a JSON
    `WsEvent` (see components). The WebSocket is not described as an operation because
    OpenAPI can't express it; its payloads are.
servers:
  - url: /api
paths:
...
  /chats:
    get:
      operationId: listChats
      tags: [chats]
      responses:
        "200":
          description: All chats, most recently active first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Chat" }
        "401": { $ref: "#/components/responses/Error" }
...
  /chats/{chatId}/messages:
    parameters:
      - $ref: "#/components/parameters/ChatId"
    get:
      operationId: listMessages
      tags: [chats]
      parameters:
        - name: before
          in: query
          required: false
          description: Return messages older than this message id
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Messages in chronological order (oldest first)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessagePage" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    post:
      operationId: sendMessage
      tags: [chats]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendMessageRequest" }
      responses:
        "201":
          description: Message created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /chats/{chatId}/read:
    parameters:
      - $ref: "#/components/parameters/ChatId"
    post:
      operationId: markChatRead
      tags: [chats]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MarkReadRequest" }
      responses:
        "204": { description: Marked as read }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      responses:
        "200":
          description: All agents
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Agent" }
        "401": { $ref: "#/components/responses/Error" }
  /agents/{agentId}:
    parameters:
      - name: agentId
        in: path
        required: true
        schema: { type: string }
    patch:
      operationId: updateAgent
      tags: [agents]
      description: Change an agent's settings. Broadcasts an agent.updated event.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateAgentRequest" }
      responses:
        "200":
          description: The updated agent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteAgent
      tags: [agents]
      description: |
        Delete an agent for good: its DM and the messages in it, memories, tasks, conversation
        history, access to connections, and its sandbox when no other agent shares it. Its
        messages in group chats stay, without an author. The last admin agent can't be deleted.
        Broadcasts agent.deleted.
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409":
          description: It's the last admin agent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /agents/{agentId}/memories:
    parameters:
      - name: agentId
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: listMemories
      tags: [agents]
      description: The durable memories an agent saved, oldest first.
      responses:
        "200":
          description: Memories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Memory" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /agents/{agentId}/memories/{memoryId}:
    parameters:
      - name: agentId
        in: path
        required: true
        schema: { type: string }
      - name: memoryId
        in: path
        required: true
        schema: { type: string }
    delete:
      operationId: deleteMemory
      tags: [agents]
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /agents/{agentId}/archive:
    parameters:
      - name: agentId
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: archiveAgent
      tags: [agents]
      description: |
        Archive an agent: it stops working and disappears from the app (its DM is hidden). The
        last admin agent can't be archived. Broadcasts agent.archived.
      responses:
        "204": { description: Archived }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409":
          description: The agent can't be archived (e.g. it's the only admin agent)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /agents/{agentId}/approvals:
    parameters:
      - name: agentId
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: listStandingApprovals
      tags: [agents]
      description: Actions the user always allows this agent to take without asking.
      responses:
        "200":
          description: The standing approvals, oldest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StandingApproval" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /agents/{agentId}/approvals/{approvalId}:
    parameters:
      - name: agentId
        in: path
        required: true
        schema: { type: string }
      - name: approvalId
        in: path
        required: true
        schema: { type: string }
    delete:
      operationId: revokeStandingApproval
      tags: [agents]
      description: Take a standing approval back, so the agent asks again.
      responses:
        "204": { description: Revoked }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /agents/{agentId}/sandbox:
    parameters:
      - name: agentId
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getSandbox
      tags: [agents]
      description: The agent's computer (Docker sandbox)
      responses:
        "200":
          description: Sandbox state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sandbox" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /agents/{agentId}/sandbox/restart:
    parameters:
      - name: agentId
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: restartSandbox
      tags: [agents]
      description: Restart the agent's computer. Running processes stop; files in /home/agent stay.
      responses:
        "200":
          description: Sandbox state after the restart
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sandbox" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /agents/{agentId}/tasks:
    parameters:
      - name: agentId
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: listTasks
      tags: [agents]
      description: The agent's scheduled tasks, oldest first.
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Task" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /tasks/{taskId}:
    parameters:
      - name: taskId
        in: path
        required: true
        schema: { type: string }
    patch:
      operationId: updateTask
      tags: [agents]
      description: Pause or resume a task.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateTaskRequest" }
      responses:
        "200":
          description: The updated task
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteTask
      tags: [agents]
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /push/config:
    get:
      operationId: getPushConfig
      tags: [settings]
      description: The VAPID public key browsers subscribe with.
      responses:
        "200":
          description: Push config
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PushConfig" }
        "401": { $ref: "#/components/responses/Error" }
  /push/subscriptions:
    post:
      operationId: subscribePush
      tags: [settings]
      description: Register this browser for notifications.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PushSubscription" }
      responses:
        "204": { description: Subscribed }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
    delete:
      operationId: unsubscribePush
      tags: [settings]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PushUnsubscribe" }
      responses:
        "204": { description: Unsubscribed }
        "401": { $ref: "#/components/responses/Error" }
...
  /settings/timezone:
    put:
      operationId: setTimezone
      tags: [settings]
      description: |
        Save the user's IANA time zone (the web app sends the browser's). Agents' schedules and
        "current time" use it.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetTimezoneRequest" }
      responses:
        "204": { description: Saved }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /agents/{agentId}/retry:
    parameters:
      - name: agentId
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: retryAgent
      tags: [agents]
      description: |
        Re-run the agent's last turn after a failure, from exactly where it stopped. The result
        arrives as usual over the WebSocket.
      responses:
        "202": { description: Retry queued }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409":
          description: The agent is busy and will pick up pending work on its own
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /messages/{messageId}/answer:
    parameters:
      - name: messageId
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: answerPrompt
      tags: [chats]
      description: |
        Answer a pending prompt (a message with `prompt` set). The answer is stored on the
        message (broadcast as message.updated) and delivered to the agent that asked.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PromptAnswer" }
      responses:
        "200":
          description: The updated message
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409":
          description: The prompt was already answered or dismissed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /models:
    get:
      operationId: listModels
      tags: [models]
      responses:
        "200":
          description: Models available from the provider
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Model" }
        "401": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: openbot_session
  parameters:
    ChatId:
      name: chatId
      in: path
      required: true
      schema: { type: string }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Chat:
      type: object
      required: [id, kind, name, members, unreadCount, lastMessage, createdAt, categoryId, position]
      properties:
        id: { type: string }
        kind: { type: string, enum: [dm, group] }
        name:
          type: string
          description: Group name, or the agent's name for DMs
        members:
          type: array
          description: Agent participants (the user is implicit in every chat)
          items: { $ref: "#/components/schemas/ChatMember" }
        unreadCount: { type: integer }
        lastMessage:
          allOf: [{ $ref: "#/components/schemas/Message" }]
          nullable: true
        createdAt: { type: string, format: date-time }
        categoryId:
          type: string
          nullable: true
          description: The sidebar category it's in; null = Unassigned
        position:
          type: integer
          nullable: true
          description: |
            Its place within its section once the user arranged it. Within a section, chats
            with no position (new or never moved) come first, most recently active first, then
            the rest by position.
...
security:
  - session: []